requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.18.0
onnx>=1.15.0
onnxruntime>=1.18.0
numpy
scikit-learn
pyyaml
tqdm
